backend/requirements.txt
fastapi
uvicorn
langchain-groq
langchain-core
python-dotenv
pydantic
